# eCommerce/ProductsService/products-service-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ProductsService.API",
    "version": "1.0"
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": [
          "ProductsService.API"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ProductsService.API"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ProductsService.API"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products/search/product-id/{ProductID}": {
      "get": {
        "tags": [
          "ProductsService.API"
        ],
        "parameters": [
          {
            "name": "ProductID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products/search/{SearchString}": {
      "get": {
        "tags": [
          "ProductsService.API"
        ],
        "parameters": [
          {
            "name": "SearchString",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/products/{ProductID}": {
      "delete": {
        "tags": [
          "ProductsService.API"
        ],
        "parameters": [
          {
            "name": "ProductID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CategoryOptions": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProductAddRequest": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/CategoryOptions"
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "quantityInStock": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductUpdateRequest": {
        "type": "object",
        "properties": {
          "productID": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/CategoryOptions"
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "quantityInStock": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}